refactor: 에러 시나리오 기반 service 테스트 코드 작성
