+++ src/test/java/com/homesweet/homesweetback/domain/community/repository/CommunityRepositoryTest.java
@@ -11,15 +11,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-// @DataJpaTest
 @SpringBootTest
 @Transactional
-class CommunityPostRepositoryTest {
+class CommunityRepositoryTest {
 
     @Autowired
     private CommunityPostRepository postRepository;
